+++ packages/@runlightyear/linear/src/teams/getTeam.ts
@@ -1,10 +1,10 @@
-import { ID } from "../types/ID";
+import { LinearID } from "../types/LinearID";
 import { HttpProxyResponse } from "@runlightyear/lightyear";
 import { Linear } from "../Linear";
 import { TeamResponse, teamResponseFields } from "./TeamResponse";
 
 export interface GetTeamProps {
-  id: ID;
+  id: LinearID;
 }
 
 const query = `