+++ packages/@runlightyear/gmail/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/gmail
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [820276f3]
+  - @runlightyear/lightyear@0.20.1
+
 ## 0.1.1
 
 ### Patch Changes